hshammog/src/requirements.txt
argparse==1.2.1
config==0.3.9
kazoo==2.0
pykafka==0.1.3
pyzmq==14.5.0
pyzmq-static==2.2
wsgiref==0.1.2
